text_to_sql_app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
google-generativeai>=0.7.0
pydantic
python-dotenv
